Fix warning
When importing the AppSync SDK into the starter app (aws-mobile-appsync-events-starter-ios), the SDK triggers a warning.

To reproduce:
- Clone aws-mobile-appsync-events-starter-ios using Xcode 9.2
- Run `pod install` (I used version 1.4.0.rc.1)
- Open EventsApp.xcworkspace and build

This produces the warning:
```
aws-mobile-appsync-events-starter-ios/Pods/AWSAppSync/AWSAppSyncClient/MQTTSDK/SocketRocket/AWSSRWebSocket.m:662:50: 'NSURLNetworkServiceTypeCallSignaling' is only available on iOS 10.0 or newer
```

In addition, this commit also assigns the correct network service type for `NSURLNetworkServiceTypeCallSignaling`.

Note: I tried to fix this problem with a compiler directive (see https://github.com/facebook/SocketRocket/pull/450/commits/bdb09bfc9e09c5e2d2d1f6734a3ffdc236ebb955), but this would trigger the warning "Enumeration value 'NSURLNetworkServiceTypeCallSignaling' not handled in switch" in the start app instead. For this reason, I solved this problem with a default case and `@available`.

Also, it looks like the official SocketRocket project removed the need for this method – might be worth to upgrate to the latest version