[#193] Suppress auto-reconnect in the low-level AWSIoTMQTTClient
This fixes the connection thrashing that can occur when the MQTT client
reconnects while AppSync is simultaneously advising clients to reconnect. The
preferred level at which to manage reconnections for AppSync is the AppSync
client itself, rather than the MQTTClient.

It's possible at some point in the future we may elect to expose an AppSync
client to support auto-reconnect, but even that would still be managed in the
AppSync client, rather than the MQTT client.